requirements.txt
asn1crypto==0.24.0
astroid==1.6.6
attrs==17.4.0
Automat==0.6.0
awscli==1.16.218
blinker==1.4
boto3==1.10.12
botocore==1.13.12
certifi==2019.9.11
chardet==3.0.4
click==6.7
colorama==0.3.9
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
dj-database-url==0.5.0
Django==1.11.26
django-multiupload==0.5.2
django-forms-bootstrap==3.1.0
django-storages==1.7.2
django-widget-tweaks==1.4.5
docutils==0.15.2
future==0.18.2
gunicorn==19.9.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.8
ikp3db==1.1.4
incremental==16.10.1
isort==4.3.21
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.4
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.4.1
MarkupSafe==1.0
mccabe==0.6.1
netifaces==0.10.4
oauthlib==2.0.6
olefile==0.45.1
parso==0.1.1
pbr==5.4.2
Pillow==5.1.0
psycopg2-binary==2.8.4
pyasn1==0.4.6
pyasn1-modules==0.2.1
pycrypto==2.6.1
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.8.1
pylint-django==0.8.0
pylint-flask==0.5
pylint-plugin-utils==0.5
pyOpenSSL==17.5.0
pyserial==3.4
python-dateutil==2.8.0
python-debian==0.1.32
python-http-client==3.2.1
pytz==2019.3
pyxdg==0.25
PyYAML==5.1
requests==2.22.0
requests-unixsocket==0.1.5
roman==2.0.0
rsa==3.4.2
s3transfer==0.2.1
SecretStorage==2.3.1
sendgrid==6.1.0
service-identity==16.0.0
six==1.13.0
sqlparse==0.3.0
stevedore==1.30.1
stripe==2.38.0
Twisted==17.9.0
urllib3==1.25.6
virtualenv==16.2.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wrapt==1.11.2
zope.interface==4.3.2
